Fix local and global resolving
- Fix issue, where completion and resolving showed symbols that came from a
completely different project. The reason was that GlobalSearchScope does
not work as one would expect. This issue is fixed now and
resolving of external symbols from other files or libraries should work
correctly.
- Fix for a minor issue in resolving patterns from Rule. In a_ -> a, the
a_ should be green while the a is a global symbol
- External Mathematica libraries are now regarded as "being compiled" to
make navigation and resolving work correctly. You need to recreate your
libraries in your project settings!
- Fix local resolving of Condition expressions
- Fix for the FullForm viewer